requirements.txt
numpy>=1.24
matplotlib>=3.7
Pillow>=9.0
biopython>=1.80        # optional: for C-optimised alignment in timing benchmark
